+++ code/game/g_local.h
@@ -428,7 +428,10 @@ struct gclient_s {
 	int			forcePowerSoundDebounce; //if > level.time, don't do certain sound events again (drain sound, absorb sound, etc)
 
 	qboolean	fjDidJump;
+
 	vec3_t		JKPlusPauseSavedView; // Tr!Force: [Pause] Player saved view angles
+	int			JKPlusTimeResidual; // Tr!Force: Player time residual
+	int			JKPlusDropFlagTimer; // Tr!Force: [DropFlag] Player timer
 };